Add day 1
I originally implemented day 1 in Ruby since it's combinations and I'm
lazy, but rewrote it in Zig because I've been meaning to make an excuse
to learn Zig.

The combination function is based off one written in nim because
I cannot for the life of me remember how to implement a combination
function as an iterator off the top of my head.