Add day 3
This one's pretty simple still. Just walks over lines and checks for
hits over a given stride. It was the first use I made of default field
values for structs, so learned something in the process.

Unlike last year (I think?), it seems like there's not a lot of
revisiting code yet. Maybe that'll change in day 4.