Add day 4 code
This one was fun. Wrote a lexer, albeit only a minimal one, to pull the
words and separators (two newlines) out of the input, since that seemed
like a better idea than trying to split and interate over things.
Probably wasn't necessary, but I like writing lexers, so this is fine.