Add day 7
Currently writing this in a house that's been almost pitch black for
several hours now due to a power outage. Also have a mild headache. Had
to submit my results via phone since tethering has been inconsistent
and only lasted long enough to get the sample input and real input
(included as input_rules.zig).

Most of the work here is in the two ruby files, each of which compiles
the source text to Zig or graphviz (the latter was just so I could
eyeball it and see if there were cycles).