Add day 8
This one just implements a simple VM to run the accumulator. The hits
array is a little unfortunate since I was being lazy. There is probably
a better way to indicate that a particular instruction was hit multiple
times, but it's not really necessary to do for this.

Part 2 involves mostly brute forcing the problem by just stepping
through the code with the VM and replacing jmp/nop instructions until
the correct one is found. When I submitted the answer, I eyeballed the
instructions after walking back through the code from the looping
point manually, so there was no code for part 2 until after submission
because it was frankly faster to do it by hand.

I'm not sure about the parseCode function. I'm not sure that it matters
to have a generalized parseCode function, but I felt like trying it.
Probably burned more time on that than the actual parsing.