Add day 9
This one was pretty boring and just involves a combination and
a sliding window. Since I had the combination type from day one, I just
reused that for part one. For part 2, the sliding window is easy enough
to do in a for loop (though it'd be fancier with an iterator-ish type).

Got to learn about std.sort, so time well spent.