Add day 10 solution
There's more commentary in the code than the commit this time because
I spent about an hour or more trying to wrap my head around the
problem, then flailed around for a while longer with combinations of
runs of 1-sequences before getting a hint on Reddit about the
tribonacci sequence being useful for this. So, part 2 was dismal.
Interesting problem, but very tiring.